+++ Plugins/ActorPools/Source/NexusActorPools/Public/NKillZoneComponent.h
@@ -25,19 +25,28 @@ public:
 	virtual void BeginPlay() override;
 	virtual void EndPlay(const EEndPlayReason::Type EndPlayReason) override;
 
+	/**
+	 * Gets the internal counter tracking the number of AActors the component has killed.
+	 * 
+	 * @return The kill count.
+	 */
 	UFUNCTION(BlueprintCallable, DisplayName="Get Kill Count", Category = "NEXUS|Actor Pools",
-		meta=(ToolTip="Gets the internal counter tracking the number of AActors the component has killed.",
-		DocsURL="https://nexus-framework.com/docs/plugins/actor-pools/types/kill-zone-component/#get-kill-count"))
+		meta=(DocsURL="https://nexus-framework.com/docs/plugins/actor-pools/types/kill-zone-component/#get-kill-count"))
 	int32 GetKillCount() const { return KillCount; }
-	
+
+	/**
+	 * Sets the internal counter tracking the number of AActors the component has killed.
+	 * @param NewKillCount The new value to use as the kill count.
+	 */
 	UFUNCTION(BlueprintCallable, DisplayName="Set Kill Count", Category = "NEXUS|Actor Pools",
-		meta=(ToolTip="Sets the internal counter tracking the number of AActors the component has killed.",
-		DocsURL="https://nexus-framework.com/docs/plugins/actor-pools/types/kill-zone-component/#set-kill-count"))
+		meta=(DocsURL="https://nexus-framework.com/docs/plugins/actor-pools/types/kill-zone-component/#set-kill-count"))
 	void SetKillCount(const int32 NewKillCount) { KillCount = NewKillCount; }
 
+	/**
+	 * Resets the internal counter tracking the number of AActors the component has killed to 0.
+	 */
 	UFUNCTION(BlueprintCallable, DisplayName="Reset Kill Count", Category = "NEXUS|Actor Pools",
-		meta=(ToolTip="Resets the internal counter tracking the number of AActors the component has killed to 0.",
-		DocsURL="https://nexus-framework.com/docs/plugins/actor-pools/types/kill-zone-component/#reset-kill-count"))
+		meta=(DocsURL="https://nexus-framework.com/docs/plugins/actor-pools/types/kill-zone-component/#reset-kill-count"))
 	void ResetKillCount() { KillCount = 0; }
 
 	UFUNCTION()
@@ -45,9 +54,7 @@ public:
 		int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepHitResult);
 	
 protected:
-	/**
-	 * Should static/station actors be ignored?
-	 */
+
 	UPROPERTY(EditInstanceOnly, BlueprintReadWrite, DisplayName="Ignore Static Actors", Category="Kill Zone")
 	bool bIgnoreStaticActors = true;